refactor getCodeOwners and move hooks from api to referrals
