+++ src/Api/referrals.js
@@ -1,9 +1,22 @@
-import { ethers } from "ethers";
+import { constants, ethers } from "ethers";
 import { gql } from "@apollo/client";
 import { useState, useEffect } from "react";
+import ReferralStorage from "../abis/ReferralStorage.json";
 
-import { ARBITRUM, AVALANCHE, MAX_REFERRAL_CODE_LENGTH, bigNumberify } from "../Helpers";
+import {
+  ARBITRUM,
+  AVALANCHE,
+  MAX_REFERRAL_CODE_LENGTH,
+  bigNumberify,
+  isAddressZero,
+  helperToast,
+  getProvider,
+  fetcher,
+} from "../Helpers";
 import { arbitrumReferralsGraphClient, avalancheReferralsGraphClient } from "./common";
+import { getContract } from "../Addresses";
+import { callContract } from ".";
+import useSWR from "swr";
 const ACTIVE_CHAINS = [ARBITRUM, AVALANCHE];
 
 function getGraphClient(chainId) {
@@ -40,32 +53,38 @@ export function encodeReferralCode(code) {
   return ethers.utils.formatBytes32String(final);
 }
 
-async function getCodeOwnersData(network, account, codes) {
-  const referralCodeOwnerQuery = (referralCode) =>
-    gql(
-      `{
-      referralCodes(where: {code: "${referralCode}"}) {
+async function getCodeOwnersData(network, account, codes = []) {
+  if (codes.length < 1 || !account || !network) {
+    return undefined;
+  }
+  const query = gql(
+    `query allCodes($codes: [String!]!){
+      referralCodes(where: {code_in: $codes}) {
         owner
+        id
       }
     }`
-    );
-
-  return Promise.all(
-    codes.map((code) => {
-      return getGraphClient(network)
-        .query({ query: referralCodeOwnerQuery(code) })
-        .then(({ data }) => {
-          const owner = data.referralCodes[0]?.owner;
-          return {
-            code,
-            codeString: decodeReferralCode(code),
-            owner,
-            isTaken: !!owner,
-            isTakenByCurrentUser: owner && String(owner).toLowerCase() === String(account).toLowerCase(),
-          };
-        });
-    })
   );
+
+  return getGraphClient(network)
+    .query({ query, variables: { codes } })
+    .then(({ data }) => {
+      const { referralCodes } = data;
+      const codeOwners = referralCodes.reduce((acc, cv) => {
+        acc[cv.id] = cv.owner;
+        return acc;
+      }, {});
+      return codes.map((code) => {
+        const owner = codeOwners[code] || constants.AddressZero;
+        return {
+          code,
+          codeString: decodeReferralCode(code),
+          owner,
+          isTaken: !isAddressZero(owner),
+          isTakenByCurrentUser: owner && String(owner).toLowerCase() === String(account).toLowerCase(),
+        };
+      });
+    });
 }
 
 export function useUserCodesOnAllChain(account) {
@@ -93,7 +112,7 @@ export function useUserCodesOnAllChain(account) {
             })
         )
       );
-      const [codeOwnersOnAvax, codeOwnersOnArbitrum] = await Promise.all([
+      const [codeOwnersOnAvax = [], codeOwnersOnArbitrum = []] = await Promise.all([
         getCodeOwnersData(AVALANCHE, account, arbitrumCodes),
         getCodeOwnersData(ARBITRUM, account, avalancheCodes),
       ]);
@@ -120,6 +139,7 @@ export function useReferralsData(chainId, account) {
   const [data, setData] = useState();
   const [loading, setLoading] = useState(true);
   const ownerOnOtherChain = useUserCodesOnAllChain(account);
+
   useEffect(() => {
     if (!chainId) return;
 
@@ -289,5 +309,72 @@ export function useReferralsData(chainId, account) {
   return {
     data: data || null,
     loading,
+  };
+}
+
+export async function registerReferralCode(chainId, referralCode, { library, ...props }) {
+  const referralStorageAddress = getContract(chainId, "ReferralStorage");
+  const contract = new ethers.Contract(referralStorageAddress, ReferralStorage.abi, library.getSigner());
+  return callContract(chainId, contract, "registerCode", [referralCode], { ...props });
+}
+export async function setTraderReferralCodeByUser(chainId, referralCode, { library, ...props }) {
+  const referralStorageAddress = getContract(chainId, "ReferralStorage");
+  const contract = new ethers.Contract(referralStorageAddress, ReferralStorage.abi, library.getSigner());
+  const codeOwner = await contract.codeOwners(referralCode);
+  if (isAddressZero(codeOwner)) {
+    helperToast.error("Referral code does not exist");
+    return new Promise((resolve, reject) => {
+      reject();
+    });
+  }
+  return callContract(chainId, contract, "setTraderReferralCodeByUser", [referralCode], {
+    ...props,
+  });
+}
+export async function getReferralCodeOwner(chainId, referralCode) {
+  const referralStorageAddress = getContract(chainId, "ReferralStorage");
+  const provider = getProvider(null, chainId);
+  const contract = new ethers.Contract(referralStorageAddress, ReferralStorage.abi, provider);
+  const codeOwner = await contract.codeOwners(referralCode);
+  return codeOwner;
+}
+
+export function useUserReferralCode(library, chainId, account) {
+  const referralStorageAddress = getContract(chainId, "ReferralStorage");
+  const { data: userReferralCode, mutate: mutateUserReferralCode } = useSWR(
+    account && [`ReferralStorage:traderReferralCodes`, chainId, referralStorageAddress, "traderReferralCodes", account],
+    {
+      fetcher: fetcher(library, ReferralStorage),
+    }
+  );
+  return {
+    userReferralCode,
+    mutateUserReferralCode,
+  };
+}
+export function useReferrerTier(library, chainId, account) {
+  const referralStorageAddress = getContract(chainId, "ReferralStorage");
+  const { data: referrerTier, mutate: mutateReferrerTier } = useSWR(
+    account && [`ReferralStorage:referrerTiers`, chainId, referralStorageAddress, "referrerTiers", account],
+    {
+      fetcher: fetcher(library, ReferralStorage),
+    }
+  );
+  return {
+    referrerTier,
+    mutateReferrerTier,
+  };
+}
+export function useCodeOwner(library, chainId, account, code) {
+  const referralStorageAddress = getContract(chainId, "ReferralStorage");
+  const { data: codeOwner, mutate: mutateCodeOwner } = useSWR(
+    account && code && [`ReferralStorage:codeOwners`, chainId, referralStorageAddress, "codeOwners", code],
+    {
+      fetcher: fetcher(library, ReferralStorage),
+    }
+  );
+  return {
+    codeOwner,
+    mutateCodeOwner,
   };
 }
